Fix two issues with Visual Studio 2008.
BUG=1290595

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@126 0039d316-1c4b-4281-b951-d872f2087c98

+++ sandbox/src/policy_engine_opcodes.cc
@@ -299,7 +299,7 @@ EvalResult OpcodeEval<OP_WSTRING_MATCH>(PolicyOpcode* opcode,
   // Advance the source string to the last successfully evaluated position
   // according to the match context.
   source_str = &source_str[context->position];
-  int source_len  = g_nt.wcslen(source_str);
+  int source_len  = static_cast<int>(g_nt.wcslen(source_str));
 
   if (0 == source_len) {
     // If we reached the end of the source string there is nothing we can
